requirements.txt
transformers==4.51.3
torch==2.6.0
sentence-transformers==4.1.0
pandas==2.2.3
datasets==3.5.0
openai==1.74.0
pydantic==2.11.3
tiktoken==0.9.0
litellm==1.67.5
duckdb==1.2.2
sql_metadata==2.15.0
python-dotenv==1.1.0
matplotlib==3.10.1